+++ src/test/nextflow/co2footprint/CO2FootprintPluginTest.groovy
@@ -79,8 +79,8 @@ class CO2FootprintPluginTest extends Specification{
     /**
      * @return A list with booleans indicating the existence of files
      */
-    List<Boolean> filesExist(Path tracePath, Path summaryPath,Path reportPath, Path provenancePath) {
-        return [tracePath, summaryPath, reportPath, provenancePath].collect({ Path path -> path.isFile() })
+    List<Boolean> filesExist(Path tracePath, Path reportPath, Path provenancePath) {
+        return [tracePath, reportPath, provenancePath].collect({ Path path -> path.isFile() })
     }
 
     /**
@@ -123,13 +123,11 @@ class CO2FootprintPluginTest extends Specification{
         when:
         Path tempPath = Files.createTempDirectory('tmpdir')
         Path tracePath = tempPath.resolve('trace_test.txt')
-        Path summaryPath = tempPath.resolve('summary_test.txt')
         Path reportPath = tempPath.resolve('report_test.html')
         Path provenancePath = tempPath.resolve('provenance_test.json')
         Map config = [
             co2footprint: [
                 'trace': ['enabled': true, 'file': tracePath],
-                'summary': ['enabled': true, 'file': summaryPath],
                 'report': ['enabled': true, 'file': reportPath],
                 'provenance': [enabled: true, file: provenancePath]
             ]
@@ -140,20 +138,18 @@ class CO2FootprintPluginTest extends Specification{
 
         then:
         observers.size() == 1
-        filesExist(tracePath, summaryPath, reportPath, provenancePath) == [true, true, true, true]
+        filesExist(tracePath, reportPath, provenancePath) == [true, true, true]
     }
 
     def 'Creation of some files'() {
         when:
         Path tempPath = Files.createTempDirectory('tmpdir')
         Path tracePath = tempPath.resolve('trace_test.txt')
-        Path summaryPath = tempPath.resolve('summary_test.txt')
         Path reportPath = tempPath.resolve('report_test.html')
         Path provenancePath = tempPath.resolve('provenance_test.json')
         Map config = [
                 co2footprint: [
                         'trace': ['enabled': true, 'file': tracePath],
-                        'summary': ['enabled': false, 'file': summaryPath],
                         'report': ['enabled': true, 'file': reportPath],
                         'provenance': [enabled: false, file: provenancePath]
                 ]
@@ -164,7 +160,7 @@ class CO2FootprintPluginTest extends Specification{
 
         then:
         observers.size() == 1
-        filesExist(tracePath, summaryPath, reportPath, provenancePath) == [true, false, true, false]
+        filesExist(tracePath, reportPath, provenancePath) == [true, true, false]
     }
 
     def 'Creation of no files'() {
@@ -174,13 +170,11 @@ class CO2FootprintPluginTest extends Specification{
         when:
         Path tempPath = Files.createTempDirectory('tmpdir')
         Path tracePath = tempPath.resolve('trace_test.txt')
-        Path summaryPath = tempPath.resolve('summary_test.txt')
         Path reportPath = tempPath.resolve('report_test.html')
         Path provenancePath = tempPath.resolve('provenance_test.json')
         Map config = [
                 co2footprint: [
                         'trace': ['enabled': false, 'file': tracePath],
-                        'summary': ['enabled': false, 'file': summaryPath],
                         'report': ['enabled': false, 'file': reportPath],
                         'provenance': [enabled: false, file: provenancePath]
                 ]
@@ -191,9 +185,9 @@ class CO2FootprintPluginTest extends Specification{
 
         then:
         observers.size() == 1
-        filesExist(tracePath, summaryPath, reportPath, provenancePath) == [false, false, false, false]
+        filesExist(tracePath, reportPath, provenancePath) == [false, false, false]
         logChecker.checkLogs(null, [
-            'No output files are enabled - to enable, set `enabled: true` in the sections `trace`, `summary` or `report`.'
+            'No output files are enabled - to enable, set `enabled: true` in the sections `trace` or `report`.'
         ])
     }
 }